app/requirements.txt
Flask
Flask-B3
Flask-Sleuth
statsd
gunicorn